DetailsMovieScreen show actors movie, rename viewmodels, items
